# FoodApp/Documentation/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FoodApp API",
    "version": "v1"
  },
  "paths": {
...
    "/api/tags": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Retrieves a list of tags with optional filtering.",
        "description": "Requires user authentication. If name is provided, tags containing that text in their name are returned.\r\nResults are sorted alphabetically by name. Tags are used to categorize recipes for easier discovery.\r\nCommon tag examples include \"vegetarian\", \"gluten-free\", \"quick-meal\", \"italian\", etc.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Optional name to search for (partial match, case-insensitive).",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the list of matching tags."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          }
        }
      },
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "Creates a new tag.",
        "description": "Requires admin role authentication. The tag name must be unique within the system.\r\nTag names are trimmed of whitespace and validated for minimum length.\r\nTags help organize recipes by categories like dietary restrictions, cooking methods,\r\ncuisine types, or any other meaningful classification system.",
        "requestBody": {
          "description": "The tag data to create. Name is required and must be unique.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the created tag."
          },
          "400": {
            "description": "Bad Request - invalid tag data or missing required fields."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "403": {
            "description": "Forbidden - user does not have admin role required for this operation."
          }
        }
      }
    },
    "/api/tags/{id}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Tag"
        ],
        "summary": "Updates an existing tag.",
        "description": "Requires admin role authentication. The tag object must have its ID property set to match the route parameter.\r\nThe updated name must be unique within the system. Returns 400 if no fields are provided for update.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the tag to update (must match the tag's ID in the request body).",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The updated tag data. The ID in the body must match the route parameter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated tag."
          },
          "400": {
            "description": "Bad Request - no valid fields provided for update or tag not found."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "403": {
            "description": "Forbidden - user does not have admin role required for this operation."
          }
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Deletes a tag by its ID.",
        "description": "Requires admin role authentication. This operation performs a cascading update:\r\n- The tag is permanently removed from the database\r\n- All recipe-tag associations for this tag are also removed\r\n- Recipes themselves remain in the system, they just lose this tag classification\r\nThis is a safe operation that preserves recipe data while removing the categorization.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the tag to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns deletion status (true if successful, false if not found)."
          },
          "401": {
            "description": "Unauthorized - valid authentication token required."
          },
          "403": {
            "description": "Forbidden - user does not have admin role required for this operation."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TagCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TagUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Put **_ONLY_** your JWT Bearer token on textbox below!",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}